Add services and controllers layers

+++ exercises/23-node-js-camada-servico-rest-restful/01-day-camada-model/02-exercises/00-content-review/02-exercise/models/userModel.js
@@ -1,6 +1,7 @@
 const connection = require('../models/connection');
 
-const formatedUsersResult = (movies) => movies.map((movie) => ({
+// QUESTION - Onde seria um bom lugar para se colocar esse tipo de função?
+const formatUsersResult = (movies) => movies.map((movie) => ({
   id: movie.id,
   firstName: movie.first_name,
   lastName: movie.last_name,
@@ -9,13 +10,13 @@ const formatedUsersResult = (movies) => movies.map((movie) => ({
 
 // Funções que fazem conexão com o banco de dados
 
-const create = async (firstName, lastName, email, password) => {
+const create = async ({ firstName, lastName, email, password }) => {
   const query = 'INSERT INTO users_crud.users (first_name, last_name, email, password) VALUES(?, ?, ?, ?)';
   const [result] = await connection.execute(query, [firstName, lastName, email, password]);
 
-  console.log('create - result:', result);
+  console.log('userModel - create - result:', result);
 
-  // Como verificar o resultado inserido????
+  // QUESTION - Como verificar o resultado inserido? Eu realmente preciso fazer isso?
   if (!result) return null;
 
   return {
@@ -31,46 +32,52 @@ const getAll = async () => {
 
   const [users] = await connection.execute(query);
 
+  // QUESTION - essa é uma boa maneira de se fazer uma verificação? Se não, como seria?
   if (!users.length) return [];
 
-  return formatedUsersResult(users);
+  return formatUsersResult(users);
 };
 
 const getById = async (id) => {
   const query = 'SELECT * FROM users_crud.users WHERE id = ?';
 
   const [user] = await connection.execute(query, [id]);
-  console.log('getById - user:', user);
+  console.log('Model - getById - user:', user);
   console.log('!user.length:', !user.length);
 
   if (!user.length) return null;
 
-  const userFormated = formatedUsersResult(user)[0];
+  const userFormated = formatUsersResult(user)[0];
 
   return userFormated;
 };
 
-const update = async (id, firstName, lastName, email, password) => {
+const update = async (id, { firstName, lastName, email, password }) => {
+  const doesUserExists = await getById(id);
+  console.log('Model - update - doesUserExists:', doesUserExists);
+
+  if (!doesUserExists) return null;
+
   const query = 'UPDATE users_crud.users SET first_name = ?, last_name = ?, email = ?, password = ? WHERE id = ?';
-
   const parsedId = parseInt(id, 10);
-  await connection.execute(query, [
-    firstName,
-    lastName,
-    email,
-    password,
-    parsedId,
-  ]);
 
-  return getById(id);
+  await connection.execute(query, [firstName, lastName, email, password, parsedId]);
+
+  const updatedUserData = await getById(id);
+
+  return updatedUserData;
 };
 
 const exclude = async (id) => {
+  const doesUserExists = await getById(id);
+
+  if (!doesUserExists) return false;
+
   const query = 'DELETE FROM users_crud.users WHERE id = ?';
-  
+
   await connection.execute(query, [id]);
 
-  return getById(id);
+  return doesUserExists;
 };
 
 module.exports = {
